+++ app/domains/power/schemas.py
@@ -1,6 +1,6 @@
 from typing import List, Optional, Literal
 from pydantic import BaseModel, ConfigDict, Field, field_validator, model_validator
-from decimal import Decimal
+
 from datetime import datetime
 
 
@@ -8,7 +8,7 @@ from datetime import datetime
 class UsvModuleBase(BaseModel):
     usv_unit_id: int
     slot: int = Field(..., ge=0)
-    leistung_kw: Decimal = Field(..., gt=0)
+    leistung_kw: float = Field(..., gt=0)
     status: str = "aktiv"
     seriennummer: Optional[str] = None
 
@@ -19,7 +19,7 @@ class UsvModuleCreate(UsvModuleBase):
 
 class UsvModuleUpdate(BaseModel):
     slot: Optional[int] = Field(None, ge=0)
-    leistung_kw: Optional[Decimal] = Field(None, gt=0)
+    leistung_kw: Optional[float] = Field(None, gt=0)  # type: ignore
     status: Optional[str] = None
     seriennummer: Optional[str] = None
 
@@ -34,11 +34,11 @@ class UsvUnitBase(BaseModel):
     bezeichnung: str
     hersteller: str = "Wöhrle SVS"
     rack_id: int
-    max_kw: Decimal = Field(..., gt=0)
+    max_kw: float = Field(..., gt=0)
     battery_strings: int = 2
     blocks_per_string: int = 32
-    block_voltage_v: Decimal = Decimal("12.00")
-    block_capacity_ah: Decimal = Decimal("100.00")
+    block_voltage_v: float = float("12.00")  # type: ignore
+    block_capacity_ah: float = float("100.00")  # type: ignore
 
 
 class UsvUnitCreate(UsvUnitBase):
@@ -49,11 +49,11 @@ class UsvUnitUpdate(BaseModel):
     bezeichnung: Optional[str] = None
     hersteller: Optional[str] = None
     rack_id: Optional[int] = None
-    max_kw: Optional[Decimal] = Field(None, gt=0)
+    max_kw: Optional[float] = Field(None, gt=0)  # type: ignore
     battery_strings: Optional[int] = None
     blocks_per_string: Optional[int] = None
-    block_voltage_v: Optional[Decimal] = None
-    block_capacity_ah: Optional[Decimal] = None
+    block_voltage_v: Optional[float] = None  # type: ignore
+    block_capacity_ah: Optional[float] = None  # type: ignore
 
 
 class UsvUnit(UsvUnitBase):
@@ -79,11 +79,11 @@ class UsvCalculationResponse(BaseModel):
     id: int
     berechnet_am: datetime
     usv_unit_id: int
-    last_kw: Decimal
-    peak_kw: Decimal
-    installiert_kw: Decimal
-    reserve_kw: Decimal
-    n1_kw: Decimal
+    last_kw: float
+    peak_kw: float
+    installiert_kw: float
+    reserve_kw: float
+    n1_kw: float
     kaltstart_ok: bool
     bemerkung: Optional[str] = None
     model_config = ConfigDict(from_attributes=True)
@@ -106,6 +106,7 @@ class PowerAuditResponse(BaseModel):
 # ──────────────────────────────────────────────────────────────
 Phase = Literal["L1", "L2", "L3"]
 
+
 class UsvSystemSetupRequest(BaseModel):
     """
     Eingabe für POST /api/v1/power/usv/setup.
@@ -116,11 +117,15 @@ class UsvSystemSetupRequest(BaseModel):
 
     # ── USV-Schrank ────────────────────────────────────────────
     usv_bezeichnung: str = Field(
-        ..., min_length=1, max_length=100,
+        ...,
+        min_length=1,
+        max_length=100,
         examples=["Wöhrle 40kW Schrank"],
     )
-    usv_kw: Decimal = Field(
-        ..., gt=0, le=2000,
+    usv_kw: float = Field(
+        ...,
+        gt=0,
+        le=2000,
         description="Schrank-Maximum in kW (UsvUnit.max_kw)",
         examples=[40.0],
     )
@@ -128,39 +133,51 @@ class UsvSystemSetupRequest(BaseModel):
     hersteller: str = Field(default="Wöhrle SVS", max_length=100)
 
     # ── Leistungsmodule ────────────────────────────────────────
-    module_kw: Decimal = Field(
-        ..., gt=0, le=500,
+    module_kw: float = Field(
+        ...,
+        gt=0,
+        le=500,
         description="Nennleistung je Modul in kW (symmetrische Auslegung lt. LB)",
         examples=[10.0],
     )
     module_count: int = Field(
-        ..., ge=1, le=12,
+        ...,
+        ge=1,
+        le=12,
         description="Anzahl aktiver Module bei Inbetriebnahme",
         examples=[4],
     )
     reserve_module_count: int = Field(
-        default=0, ge=0, le=12,
+        default=0,
+        ge=0,
+        le=12,
         description="Zusätzliche Module mit Status 'reserve' (N+1-Redundanz)",
     )
 
     # ── Unterverteilung / Stromkreise ──────────────────────────
     panel_bezeichnung: str = Field(
-        ..., min_length=1, max_length=100,
+        ...,
+        min_length=1,
+        max_length=100,
         examples=["UV-RZ-01"],
     )
     phases: list[Phase] = Field(
-        default_factory=lambda: ["L1", "L2", "L3"],
+        default_factory=lambda: ["L1", "L2", "L3"],  # type: ignore
         description="Phasen, für die je ein Stromkreis erzeugt wird",
     )
-    absicherung_a: Decimal = Field(
-        ..., gt=0, le=630,
+    absicherung_a: float = Field(
+        ...,
+        gt=0,
+        le=630,
         description="Absicherung je Stromkreis in Ampere",
         examples=[16.0],
     )
 
     # ── Audit (Punkt 6) ────────────────────────────────────────
     geaendert_von: str = Field(
-        ..., min_length=1, max_length=100,
+        ...,
+        min_length=1,
+        max_length=100,
         description="Freitext-Audit: wer legt das System an",
     )
 
@@ -178,14 +195,13 @@ class UsvSystemSetupRequest(BaseModel):
     def _slots_fit(self) -> "UsvSystemSetupRequest":
         total = self.module_count + self.reserve_module_count
         if total > 12:
-            raise ValueError(
-                f"Gesamtzahl Module ({total}) übersteigt max. 12 Slots."
-            )
+            raise ValueError(f"Gesamtzahl Module ({total}) übersteigt max. 12 Slots.")
         return self
 
 
 class PowerMetrics(BaseModel):
     """Ergebnis der geretteten Berechnung (siehe metrics.calculate_power_metrics)."""
+
     last_kw: float
     peak_kw: float
     installiert_kw: float
@@ -199,6 +215,7 @@ class PowerMetrics(BaseModel):
 
 class UsvSystemSetupResponse(BaseModel):
     """Was beim Setup erzeugt wurde + abgeleitete Kennzahlen + Hinweise."""
+
     usv_unit_id: int
     module_ids: list[int]
     panel_id: int
